requirements.txt
PyQt5~=5.15.9
keyboard~=0.13.5
numpy~=1.25.2
Pillow~=10.0.0
pynput~=1.7.6
opencv-python~=4.8.0.76